requirements.txt
Pillow
pysdl2
